requirements.txt
anyio==3.6.1
atomicwrites==1.4.0
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
iniconfig==1.1.1
minio==7.1.9
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pydantic==1.9.1
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
python-multipart==0.0.5
requests==2.28.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.18.1
